backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.1
PyMySQL==1.1.0
python-dotenv==1.0.0
psutil==5.9.5
cryptography==41.0.7
pytz==2023.3
